MAINT: Move ModuleDeprecationWarning and ModuleDeprecationWarning
Both should now live in the "exceptions" module.
