+++ numpy/__init__.py
@@ -106,11 +106,10 @@ Exceptions to this rule are documented.
 import sys
 import warnings
 
-from ._globals import (
-    ModuleDeprecationWarning, VisibleDeprecationWarning,
-    _NoValue, _CopyMode
-)
-from .exceptions import ComplexWarning, TooHardError, AxisError
+from ._globals import _NoValue, _CopyMode
+from .exceptions import (
+    ComplexWarning, ModuleDeprecationWarning, VisibleDeprecationWarning,
+    TooHardError, AxisError)
 
 # We first need to detect if we're being called as part of the numpy setup
 # procedure itself in a reliable manner.
